+++ src/main/java/com/benaya/link_validator_api/repository/DomainRepository.java
@@ -2,10 +2,14 @@ package com.benaya.link_validator_api.repository;
 
 import com.benaya.link_validator_api.model.Domain;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
-public interface MongodbRepository extends MongoRepository<Domain, String> {
+import java.util.Optional;
+
+@Repository
+public interface DomainRepository extends MongoRepository<Domain, String> {
     boolean existsByName(String domainName);
     Domain getByName(String domainName);
-
+    Optional<Domain> findFirstBy();
     boolean existsByUrlsContains(String url);
 }